Add files via upload
First GTK3 application in c!
This is a C program that uses the GTK+ library to create a graphical user interface. It creates a window with a toolbar and three toggle buttons, each of which opens a new window when clicked. The three windows are identified by their title ("window from button 1", "window from button 2", "window from button 3") and are destroyed when closed.

The main function initializes GTK, creates the main window, sets its properties, and connects the "destroy" signal to the gtk_main_quit function to exit the program when the main window is closed. It then creates a vertical box container and a toolbar, and adds the toolbar to the box container.

For each toggle button, it creates an icon image and a toggle button widget with a label and sets the icon widget to the image. It then connects the "clicked" signal to a callback function that creates or destroys a window depending on whether the window is already open.

The on_window_destroy, on_window_destroy1, and on_window_destroy2 functions are callback functions that are called when the "destroy" signal is emitted for each of the three windows. They set the corresponding toggle button to be inactive (unpressed) and set the window pointer to NULL.

Overall, this program creates a simple graphical interface with toggle buttons that open new windows.